Relocate LatticeLuxonUtils
